Share more internals
