+++ crates/rune/src/query/query.rs
@@ -7,8 +7,10 @@ use crate::no_std::sync::Arc;
 
 use crate::ast;
 use crate::ast::{Span, Spanned};
+use crate::compile::context::ContextMeta;
 use crate::compile::ir;
 use crate::compile::meta;
+use crate::compile::v1::GenericsParameters;
 use crate::compile::{
     self, CompileErrorKind, CompileVisitor, ComponentRef, Doc, ImportStep, IntoComponent, IrBudget,
     IrCompiler, IrInterpreter, Item, ItemBuf, ItemId, ItemMeta, Location, ModId, ModMeta, Names,
@@ -22,6 +24,12 @@ use crate::query::{Build, BuildEntry, BuiltInMacro, ConstFn, Named, QueryPath, U
 use crate::runtime::Call;
 use crate::shared::{Consts, Gen, Items};
 use crate::{Context, Hash, SourceId, Sources};
+
+enum ContextMatch<'this, 'm> {
+    Context(&'m ContextMeta, Hash),
+    Meta(&'this meta::Meta),
+    None,
+}
 
 /// The permitted number of import recursions when constructing a path.
 const IMPORT_RECURSION_LIMIT: usize = 128;
@@ -131,6 +139,193 @@ impl<'a> Query<'a> {
     /// engine.
     pub(crate) fn next_build_entry(&mut self) -> Option<BuildEntry> {
         self.inner.queue.pop_front()
+    }
+
+    // Pick private metadata to compile for the item.
+    fn select_context_meta<'this, 'm>(
+        &'this self,
+        item: ItemId,
+        metas: impl Iterator<Item = &'m ContextMeta> + Clone,
+        parameters: &GenericsParameters,
+    ) -> Result<ContextMatch<'this, 'm>, Box<QueryErrorKind>> {
+        #[derive(Debug, PartialEq, Eq, Clone, Copy)]
+        enum Kind {
+            None,
+            Type,
+            Function,
+            AssociatedFunction,
+        }
+
+        /// Determine how the collection of generic parameters applies to the
+        /// returned context meta.
+        fn determine_kind<'m>(metas: impl Iterator<Item = &'m ContextMeta>) -> Option<Kind> {
+            let mut kind = Kind::None;
+
+            for meta in metas {
+                let alt = match &meta.kind {
+                    meta::Kind::Enum { .. }
+                    | meta::Kind::Struct { .. }
+                    | meta::Kind::Type { .. } => Kind::Type,
+                    meta::Kind::Function { .. } => Kind::Function,
+                    meta::Kind::AssociatedFunction { .. } => Kind::AssociatedFunction,
+                    _ => {
+                        continue;
+                    }
+                };
+
+                if matches!(kind, Kind::None) {
+                    kind = alt;
+                    continue;
+                }
+
+                if kind != alt {
+                    return None;
+                }
+            }
+
+            Some(kind)
+        }
+
+        fn build_parameters(kind: Kind, p: &GenericsParameters) -> Option<Hash> {
+            let hash = match (kind, p.trailing, p.parameters) {
+                (_, 0, _) => Hash::EMPTY,
+                (Kind::Type, 1, [Some(ty), None]) => Hash::EMPTY.with_type_parameters(ty),
+                (Kind::Function, 1, [Some(f), None]) => Hash::EMPTY.with_function_parameters(f),
+                (Kind::AssociatedFunction, 1, [Some(f), None]) => {
+                    Hash::EMPTY.with_function_parameters(f)
+                }
+                (Kind::AssociatedFunction, 2, [Some(ty), f]) => Hash::EMPTY
+                    .with_type_parameters(ty)
+                    .with_function_parameters(f.unwrap_or(Hash::EMPTY)),
+                _ => {
+                    return None;
+                }
+            };
+
+            Some(hash)
+        }
+
+        if let Some(parameters) =
+            determine_kind(metas.clone()).and_then(|kind| build_parameters(kind, parameters))
+        {
+            if let Some(meta) = self.get_meta(item, parameters) {
+                return Ok(ContextMatch::Meta(meta));
+            }
+
+            // If there is a single item matching the specified generic hash, pick
+            // it.
+            let mut it = metas
+                .clone()
+                .filter(|i| !matches!(i.kind, meta::Kind::Macro | meta::Kind::Module))
+                .filter(|i| i.kind.as_parameters() == parameters);
+
+            if let Some(meta) = it.next() {
+                if it.next().is_none() {
+                    return Ok(ContextMatch::Context(meta, parameters));
+                }
+            } else {
+                return Ok(ContextMatch::None);
+            }
+        }
+
+        if metas.clone().next().is_none() {
+            return Ok(ContextMatch::None);
+        }
+
+        Err(Box::new(QueryErrorKind::AmbiguousContextItem {
+            item: self.pool.item(item).to_owned(),
+            infos: metas.map(|i| i.info()).collect(),
+        }))
+    }
+
+    /// Access the meta for the given language item.
+    pub(crate) fn try_lookup_meta(
+        &mut self,
+        location: Location,
+        item: ItemId,
+        parameters: &GenericsParameters,
+    ) -> compile::Result<Option<meta::Meta>> {
+        tracing::trace!("lookup meta: {:?}", item);
+
+        if parameters.is_empty() {
+            if let Some(meta) = self.query_meta(location.span, item, Default::default())? {
+                tracing::trace!("found in query: {:?}", meta);
+                self.visitor
+                    .visit_meta(location, meta.as_meta_ref(self.pool));
+                return Ok(Some(meta));
+            }
+        }
+
+        let Some(metas) = self.context.lookup_meta(self.pool.item(item)) else {
+            return Ok(None);
+        };
+
+        let (meta, parameters) = match self
+            .select_context_meta(item, metas, parameters)
+            .with_span(location.span)?
+        {
+            ContextMatch::None => return Ok(None),
+            ContextMatch::Meta(meta) => return Ok(Some(meta.clone())),
+            ContextMatch::Context(meta, parameters) => (meta, parameters),
+        };
+
+        let Some(item) = &meta.item else {
+            return Err(compile::Error::new(location.span,
+            QueryErrorKind::MissingItem {
+                hash: meta.hash,
+            }));
+        };
+
+        let meta = meta::Meta {
+            context: true,
+            hash: meta.hash,
+            item_meta: ItemMeta {
+                id: Default::default(),
+                location: Default::default(),
+                item: self.pool.alloc_item(item),
+                visibility: Default::default(),
+                module: Default::default(),
+            },
+            kind: meta.kind.clone(),
+            source: None,
+            parameters,
+        };
+
+        self.insert_meta(meta.clone()).with_span(location.span)?;
+
+        tracing::trace!("Found in context: {:?}", meta);
+
+        self.visitor
+            .visit_meta(location, meta.as_meta_ref(self.pool));
+
+        Ok(Some(meta))
+    }
+
+    /// Access the meta for the given language item.
+    pub(crate) fn lookup_meta(
+        &mut self,
+        location: Location,
+        item: ItemId,
+        parameters: impl AsRef<GenericsParameters>,
+    ) -> compile::Result<meta::Meta> {
+        let parameters = parameters.as_ref();
+
+        if let Some(meta) = self.try_lookup_meta(location, item, parameters)? {
+            return Ok(meta);
+        }
+
+        let kind = if !parameters.parameters.is_empty() {
+            CompileErrorKind::MissingItemParameters {
+                item: self.pool.item(item).to_owned(),
+                parameters: parameters.as_boxed(),
+            }
+        } else {
+            CompileErrorKind::MissingItem {
+                item: self.pool.item(item).to_owned(),
+            }
+        };
+
+        Err(compile::Error::new(location.span, kind))
     }
 
     /// Insert path information.
